Snacks App - Add/Delete to Favourite - Function Approach
Previous approach was Class Views approach, as we don't want to display this form in a different webpage; BUT instead we want ('cahier des charges' == Project Specs) to update the database on the fly in real time without changing webpage... Testing a Function approach.
Django ORM code is sound.
Need to incorporate all this in the views and interact with the 'save to favourite button'.
Then do the same to delete from favourite